Use latest cmd_lib version including our required fix
The latest version of cmd_lib's run_fun doesn't accept cmd as strings
and so we are now using the work around suggested here: https://github.com/rust-shell-script/rust_cmd_lib/discussions/33
because we don't mind injection obviously. The downside is we execute all commands through bash and not the user shell.